requirements.txt
openai
langchain-openai
python-dotenv
flask
ngrok==1.3.0
twilio